Day 7/requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
python-dotenv==1.0.1
PyPDF2==3.0.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
google-generativeai==0.5.0 